Music and switches
Fixed an issue with some switching not opening doors, added sector crush support to vertical collision detection, meaning doors are now collidable properly, we can now extract a mus file and convert it to midi for playback.

+++ DopeFish/scripts/DE_glbspProcesses/DE_glbspProcesses.gml
@@ -6,7 +6,8 @@ function DE_glbspPreProcess( sourceWad ){
 	
 	var __drive = string_copy(game_save_id,1,1);
 	
-	var __path = string_replace(game_save_id,"/","//");
+	var __path = string_replace_all(game_save_id,"\\","//");
+	//__path = string_replace_all(__path,"/","\\");
 	
 	var str = "chdir /"+__drive+" "+__path+"\nglbsp.exe "+__path+"tmp1.wad -o "+__path+"tmp2.wad";
 	str += "\nping 127.0.0.1 -n 2 > nul";
@@ -30,9 +31,11 @@ function DE_glbspProcessWad(sourceWad, destWad){
 	do{
 	}until file_exists(game_save_id+"wadBuildGLNodes.bat");
 	
-	var __tmp = external_define( "DE_GLBSP.dll", "wad_GLBSPProcess", dll_cdecl, ty_real, 1, ty_string );
+	var __tmp = external_define( game_save_id + "DE_GLBSP.dll", "wad_GLBSPProcess", dll_cdecl, ty_real, 1, ty_string );
 	
-	external_call( __tmp, game_save_id + "wadBuildGLNodes.bat" );
+	var result = external_call( __tmp, game_save_id + "wadBuildGLNodes.bat" );
+	
+	DEtrace("Creating GL Data",result);
 	
 	do{
 	}until file_exists(game_save_id+"tmp2.~wa");
@@ -64,60 +67,69 @@ function DE_glbspProcessWad(sourceWad, destWad){
 function DE_musLumpPreProcess( ){
 	
 	var file = file_text_open_write(game_save_id+"wadBuildGLNodes.bat");
-		
-	var __drive = string_copy(game_save_id,1,1);
 	
-	var __path = string_replace(game_save_id,"/","//");
+	//var __drive = string_copy(game_save_id,1,1);
 	
-	var str =__path+"\wildmidi.exe "+"tmp.mus -x "+"tmp.midi";
+	var __path = string_replace_all(game_save_id,"\\","//");
+	//__path = string_replace_all(__path,"/","\\");
+	
+	var str = "chdir "+__path+"\n"+"wildmidi.exe "+__path+"tmp.mus -x "+__path+"tmp.midi";
 	
 	file_text_write_string(file,str);
 	file_text_close(file);
 	
+	return str;
+	
 	
 }
 
-function DE_musLumpProcess( muslump ){
-	
-	DE_musLumpPreProcess();
-	
-	var lpos = DE_getLumpOfs( muslump );
+function DE_musLumpExtractToFile( lmp, f){
+	var lpos = DE_getLumpOfs( lmp );
 
-	var len = ds_list_find_value_fixed(wadDirectory, DE_getLumpNum( muslump ) ).size;
+	var len = ds_list_find_value_fixed(wadDirectory, DE_getLumpNum( lmp ) ).size;
 
 	var __tmp = buffer_create(len,buffer_fixed,1);
 
 	buffer_copy(wadbuff,lpos,len,__tmp,0);
 	
-	buffer_save(__tmp,game_save_id + "tmp.mus" );
+	buffer_save(__tmp,f );
+	
+	buffer_delete( __tmp );
+}
+
+function DE_musLumpProcess( muslump ){
+	
+	DEtrace("Pre Processing Music Lump");
+	var str = DE_musLumpPreProcess();
+	
+	do{
+	}until file_exists(game_save_id+"wadBuildGLNodes.bat");
+	
+	var file = game_save_id+"tmp.mus";
+	
+	DEtrace("Extracting Music Lump");
+	DE_musLumpExtractToFile(muslump,file);
+	
+	do{
+	}until file_exists(file);
 	
 	var __tmp = external_define( "DE_GLBSP.dll", "wad_GLBSPProcess", dll_cdecl, ty_real, 1, ty_string );
 	
-	external_call( __tmp, game_save_id + "wadBuildGLNodes.bat" );
+	var result = external_call( __tmp, game_save_id+"wadBuildGLNodes.bat" );
 	
+	DEtrace("Converting Mus to Midi",result);
+	
+	var num = 0;
 	do{
-	}until file_exists(game_save_id+"tmp.midi");
-	
-	var _mid = read_midi(game_save_id+"tmp.midi");
-	
-	noteslist = _mid[0];
-	var events = _mid[1];
-
-	var bpm = 60; //in case bpm is never declared
-	for(var i = 0; i < ds_list_size(events); i++) {
-		var event = ds_list_find_value(events, i);
-		if (event[1] == "bpm change") {
-			bpm = event[2];
-			break;
+		num ++;
+		if file_exists(game_save_id+"tmp.midi"){
+			DEtrace("Done Processing Music Lump");
+		
+			clipboard_set_text(game_save_id);
+			file_delete(file);
+			//file_delete(game_save_id+"tmp.midi");
+			file_delete(game_save_id+"wadBuildGLNodes.bat");
 		}
-	}
-
-	time = 0;
-	realbpm = 4*bpm/room_speed; //4 is an arbitrary value
-	
-	clipboard_set_text(game_save_id);
-	file_delete(game_save_id+"tmp.mus");
-	file_delete(game_save_id+"tmp.midi");
-	file_delete(game_save_id+"wadBuildGLNodes.bat");
+	}until file_exists(game_save_id+"tmp.midi") || num > 100000;
 
 }
